fix: focus imperative api is defined twice
Noticed that the focus imperative API is created twice, once here and
once in `useInput` (which already existed prior to this change). Code
was added in https://github.com/Neovici/cosmoz-input/pull/85 and it
mentions a failure in acceptance tests, but I fail to see how this
change had anything to do with the failure getting resolved.